+++ lib/classes/class-WPC_Bootstrap.php
@@ -86,6 +86,59 @@ namespace UsabilityDynamics\WPC {
        */
       public function deactivate() {}
 
+      /**
+       * Run Install Process.
+       *
+       * @author peshkov@UD
+       */
+      public function run_install_process() {
+        /* Compatibility with WP-CRM 0.36.5 and less versions */
+        $old_version = get_option( 'wp_crm_version' );
+        if( $old_version ) {
+          $this->run_upgrade_process();
+        }
+      }
+
+      /**
+       * Run Upgrade Process:
+       * - do WP-Invoice settings backup.
+       *
+       * @author peshkov@UD
+       */
+      public function run_upgrade_process() {
+        /* Do automatic Settings backup! */
+        $settings = get_option( 'wp_crm_settings' );
+
+        if( !empty( $settings ) ) {
+
+          /**
+           * Fixes allowed mime types for adding download files on Edit Product page.
+           *
+           * @see https://wordpress.org/support/topic/2310-download-file_type-missing-in-variations-filters-exe?replies=5
+           * @author peshkov@UD
+           */
+          add_filter( 'upload_mimes', function( $t ){
+            if( !isset( $t['json'] ) ) {
+              $t['json'] = 'application/json';
+            }
+            return $t;
+          }, 99 );
+
+          $filename = md5( 'wp_crm_settings_backup' ) . '.json';
+          $upload = @wp_upload_bits( $filename, null, json_encode( $settings ) );
+
+          if( !empty( $upload ) && empty( $upload[ 'error' ] ) ) {
+            if( isset( $upload[ 'error' ] ) ) unset( $upload[ 'error' ] );
+            $upload[ 'version' ] = $this->old_version;
+            $upload[ 'time' ] = time();
+            update_option( 'wp_crm_settings_backup', $upload );
+          }
+
+        }
+
+        do_action( $this->slug . '::upgrade', $this->old_version, $this->args[ 'version' ], $this );
+      }
+
     }
 
   }