+++ changes.md
@@ -1,6 +1,7 @@
 #### 1.0.0
 * Changed plugin initialization functionality.
 * Added Composer ( dependency manager ) modules and moved some functionality to composer modules ( vendors ).
+* Added doing WP-CRM Settings backup on upgrade to new version. Get information about backup: get_option('wp_crm_settings_backup');
 * Moved premium features to separate plugins.
 * Cleaned up functionality of plugin.
 * Refactored file structure of plugin.